Rename: component camel case
컴포넌트명 수정
